add change quantity in storage

+++ ManagementSystem.Common/Entities/ProductStorage.cs
@@ -14,7 +14,7 @@ namespace ManagementSystem.Common.Entities
         public int Id { get; set; }
         public int ProductId { get; set; }
         public int StorageId { get; set; }
-        public int Quantity { get; set; } = 0;
+        public float Quantity { get; set; } = 0;
         public Product Product { get; set; }
         public Storage Storage { get; set; }
     }
